+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Added
 * Oklab color model
+* XYZ color model
 
 ## [0.13.2] - 2022-06-01
 ### Added